+++ native/cocos/renderer/gfx-base/GFXBuffer.h
@@ -35,7 +35,7 @@ public:
     Buffer();
     ~Buffer() override;
 
-    static uint32_t computeHash(const BufferInfo &info);
+    static size_t computeHash(const BufferInfo &info);
 
     void initialize(const BufferInfo &info);
     void initialize(const BufferViewInfo &info);
