+++ native/cocos/renderer/gfx-gles3/states/GLES3Sampler.cpp
@@ -30,7 +30,7 @@
 namespace cc {
 namespace gfx {
 
-GLES3Sampler::GLES3Sampler(const SamplerInfo &info, uint32_t hash) : Sampler(info, hash) {
+GLES3Sampler::GLES3Sampler(const SamplerInfo &info, size_t hash) : Sampler(info, hash) {
     _typedID = generateObjectID<decltype(this)>();
 
     _gpuSampler            = CC_NEW(GLES3GPUSampler);
