+++ native/cocos/renderer/gfx-metal/MTLDevice.h
@@ -106,9 +106,9 @@ protected:
     DescriptorSetLayout *createDescriptorSetLayout() override;
     PipelineLayout *createPipelineLayout() override;
     PipelineState *createPipelineState() override;
-    GlobalBarrier *createGlobalBarrier(const GlobalBarrierInfo &info, uint32_t hash) override;
-    TextureBarrier *createTextureBarrier(const TextureBarrierInfo& info, uint32_t hash) override;
-    Sampler *createSampler(const SamplerInfo& info, uint32_t hash) override;
+    GlobalBarrier *createGlobalBarrier(const GlobalBarrierInfo &info, size_t hash) override;
+    TextureBarrier *createTextureBarrier(const TextureBarrierInfo& info, size_t hash) override;
+    Sampler *createSampler(const SamplerInfo& info, size_t hash) override;
     Swapchain *createSwapchain() override;
     void copyBuffersToTexture(const uint8_t *const *buffers, Texture *dst, const BufferTextureCopy *regions, uint count) override;
     void copyTextureToBuffers(Texture *src, uint8_t *const *buffers, const BufferTextureCopy *region, uint count) override;
