+++ native/cocos/renderer/gfx-vulkan/states/VKSampler.cpp
@@ -30,7 +30,7 @@
 namespace cc {
 namespace gfx {
 
-CCVKSampler::CCVKSampler(const SamplerInfo &info, uint32_t hash) : Sampler(info, hash) {
+CCVKSampler::CCVKSampler(const SamplerInfo &info, size_t hash) : Sampler(info, hash) {
     _typedID = generateObjectID<decltype(this)>();
 
     _gpuSampler                = CC_NEW(CCVKGPUSampler);
