+++ native/cocos/renderer/pipeline/shadow/ShadowFlow.h
@@ -62,7 +62,7 @@ private:
     vector<const scene::Light *> _validLights;
     vector<gfx::Texture *>       _usedTextures;
 
-    static std::unordered_map<uint, cc::gfx::RenderPass*> renderPassHashMap;
+    static std::unordered_map<size_t, cc::gfx::RenderPass *> renderPassHashMap;
 };
 } // namespace pipeline
 } // namespace cc
